config the deepspeed

+++ train_img_re.py
@@ -738,7 +738,7 @@ def train(epoch, model, trn_loader):
         #   compute z = f(x)
         #   maximize log p(x) = log p(z) - log |det df/dx|
 
-        x = x.to(device)
+        x = x.half().to(device)
 
         beta = beta = min(1, global_itr /
                           args.annealing_iters) if args.annealing_iters > 0 else 1.
@@ -754,7 +754,7 @@ def train(epoch, model, trn_loader):
             secmom_meter.update(secmom)
 
         if args.task in ['classification', 'hybrid']:
-            y = y.to(device)
+            y = y.half().to(device)
             crossent = criterion(logits, y)
             ce_meter.update(crossent.item())
 
@@ -851,12 +851,12 @@ def validate(epoch, model, dat_loader, phase, ema=None):
     start = time.time()
     with torch.no_grad():
         for i, (x, y) in enumerate(tqdm(dat_loader)):
-            x = x.to(device)
+            x = x.half().to(device)
             bpd, logits, _, _ = compute_loss(x, model)
             bpd_meter.update(bpd.item(), x.size(0))
 
             if args.task in ['classification', 'hybrid']:
-                y = y.to(device)
+                y = y.half().to(device)
                 loss = criterion(logits, y)
                 ce_meter.update(loss.item(), x.size(0))
                 _, predicted = logits.max(1)
